Split Pedante tests by scenario
Move shared Pedante payload builders and the fake LLM port into
support helpers, then split validation, parsing, and evaluator
integration coverage into focused test modules.

Remove the file-level `too-many-lines` suppression from the deleted
monolithic test module.
